+++ tests/pools/utils.test.mjs
@@ -1,11 +1,29 @@
+import { Worker as ThreadWorker } from 'node:worker_threads'
+import { Worker as ClusterWorker } from 'node:cluster'
 import { expect } from 'expect'
 import {
   CircularArray,
   DEFAULT_CIRCULAR_ARRAY_SIZE
 } from '../../lib/circular-array.js'
-import { updateMeasurementStatistics } from '../../lib/pools/utils.js'
+import {
+  createWorker,
+  getDefaultTasksQueueOptions,
+  updateMeasurementStatistics
+} from '../../lib/pools/utils.js'
+import { WorkerTypes } from '../../lib/index.js'
 
 describe('Pool utils test suite', () => {
+  it('Verify getDefaultTasksQueueOptions() behavior', () => {
+    const poolMaxSize = 4
+    expect(getDefaultTasksQueueOptions(poolMaxSize)).toStrictEqual({
+      concurrency: 1,
+      size: Math.pow(poolMaxSize, 2),
+      taskStealing: true,
+      tasksStealingOnBackPressure: true,
+      tasksFinishedTimeout: 2000
+    })
+  })
+
   it('Verify updateMeasurementStatistics() behavior', () => {
     const measurementStatistics = {
       history: new CircularArray()
@@ -91,5 +109,20 @@ describe('Pool utils test suite', () => {
         0.01
       )
     })
+  })
+
+  it('Verify createWorker() behavior', () => {
+    expect(
+      createWorker(
+        WorkerTypes.thread,
+        './tests/worker-files/thread/testWorker.mjs'
+      )
+    ).toBeInstanceOf(ThreadWorker)
+    expect(
+      createWorker(
+        WorkerTypes.cluster,
+        './tests/worker-files/cluster/testWorker.mjs'
+      )
+    ).toBeInstanceOf(ClusterWorker)
   })
 })